Move Concoction.ConcoctionType to ConcoctionDatabase.ConcoctionType
